Move SessionState to separate file
